requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0 